Add tests for behavior after process death
Since we do not test the real process death scenario, we instead just
manually populate the SavedStateHandle and test the behavior with that
as a fact.

The reason we do not test this in a more robust way is noted in a
comment in the test code itself too:
Note that with the SavedStateHandle.saveable APIs, the value is only
saved into the handle when there is a real event of process death which
instructs viewmodel-savedstate to trigger the save state. This is
either not yet possible to do in unit tests, or is quite hard to get
right, so we need to hack it a little but by manually populating the
SavedStateHandle ourselves with the right values, simulating what would
happen in a process death scenario.
When that artifact goes KMP it will most likely provide a much easier
way to integrate process death scenarios into our tests without having
to introduce instrumented tests.