Make background dark color only fade in instead of sliding
The entire content can now fade in and out so that the dark color is
always filling the entire screen. While the contents can still slide in
with the same transition spec as they were doing before. The important
difference being that this modifier is only added to the content in
front of the dark background here.

Also ensure that the focus is cleared on either of the two buttons
pressed with FocusManager.clearFocus()