MyInfoDestination w/o ValidationResult and with usability improvements
Now only does the necessary network request depending on which one of
the two (or both) of email and phone number have changed
The save button only appears if there was a delta between the existing
"saved" information and the new input. Instead of just appearing after
there was any change, and still being there if the member reverted back
to the real values.