Revert using SSH in VM to get the route
In the SSH side, we also need to pass in the typeMap again, and this is
going to make our approach inconsistent. If we determine this to be a
better approach we can look into it in a future PR where we do the same
thing across the entire codebase.

This reverts commit 0ea93f95