Scroll to the latest item
LatestChatMessage comes from the DB directly and won't be null even if
we've scrolled up by a lot. So we need to listen to that to be notified
about the new message, and then try to get the item at index 0 from the
pager to find the real item to scroll to. That will be null initially
when we've scrolled away from its page, so we wait a bit for it to
appear. When it does we try to scroll to it if it fits out requirements