On refresh, do not clear fresh unsent messages
We still want to be able to let our members retry to send messages that
failed to be sent recently.
This comes up very specifically in the scenario where one starts a new
conversation, they try to send more than 1 message and they all fail.
Then when retrying and succeeding, we create a new Pager from scratch,
so we refresh our RemoteMediator, which normally clears the database.
By not clearing specifically the new unsent messages we solve
accidentally deleting all those unsent messages.