Make GetAllConversationsUseCase return the right hasNewMessages status
Combines the backend information with the local DB information on the
latest seen message timestamp.
Does not mark unseen messages if the DB does not have any timestamp at
all. This should help with the migration scenario from the old chat to
the new one, where we do not want to mark all conversations as unseen
when the feature flag is turned on.