Show the "new message" tooltip at home when we get a new message
Only show it if there was no red dot initially, and a new one comes in.
If we already knew that there was a new message on the first frame do
not show it, as then it would show every single time you end up in the
home screen and there was an unread message, including when you just
start the app.