Fix retrying messages not flickering in the chat
Now the item persists in he DB while we're retrying. If it fails again,
it gets added back into the DB replacing the old entry, so the UI always
has something to show.
If it succeeds, we insert the new item in the DB, and then we delete the
old failed entry from the DB.
If we could generate the ID on the client we could do the same without
having to explicitly delete the entry locally, but we could rely on the
insert function replacing the old one with the same ID. But this works
for now too as the DB deletion happens fast enough to not be visible on
the UI.