Model the LinkInfo in one object
Now we can try to only represent valid states instead of perhaps
accidentally exposing a null link with a present buttonText.
We still fallback in the case where the backend may wrongly give us an
empty url to not show the button at all. While also falling back to a
default string if the backend decides to give a not null but empty
String accidentally.