Make all dd attribute sources reactive instead of individual coroutines
This works around the issue over at
https://github.com/DataDog/dd-sdk-android/issues/2132 which does not let
one call setUserInfo without having to also clear all previously set
attributes.
With this change, we instead keep a flow for all of our source of truth,
and combine them together so if any single one of them change, we
re-apply all of them together with one `setUserInfo` call.
We do the same for RUM too, all inside DatadogAttributesManager