Adjust ChatNotificationSender consider the CBM feature flag
Use the old link when it's off. Use a separate inbox or a conversation
specific link depending on if the notification itself contains the
conversationId in there or not.

Also replace the TaskStackBuilder with a simple
PendingIntent.getActivity since we simply have one activity to open, and
we pass it a deep link instead.
Inspiration from:
https://github.com/android/nowinandroid/blob/85129e4660f7a27c7081f4ac21169d19db89fbb6/core/notifications/src/main/kotlin/com/google/samples/apps/nowinandroid/core/notifications/SystemTrayNotifier.kt#L154-L166