Logout when receiving the unathenticated error type
This happens when we somehow are not aware that a token we possess is
actually not valid despite us thinking that it has not expired yet.
All of our services inform us about this piece of information so we can
use it to eagerly logout, which will invalidate the tokens and force the
member back into the login screen.