Introduce and use a HedvigTextButton in the DS
This also allows us to provide the implementation using our DS for the
button inside the date picker, without having to fall back to using the
material3 implementation which might not be what we want.
The "Ghost" button fits perfectly for this since it is also without a
background and this component ends up having a sensible size and
internal padding this way.