Make tooltip only take up the space it physically needs
Still keeps the touch size according to minimumInteractiveComponentSize
as far as the onClick modifier is concerned.

Also adjust the beak arrows to just be the top arrow but rotated by 90
degrees enough times to fit all scenarios.

Move the "offset" arrows to be 16.dp away from the edge they are
touching, as the design looks to have them placed

Give the tooltip a sane MIN width too, so that if there is a combination
 of a small tooltip with an off-center arrow, the arrow path still looks
  to be attached to the main squircle without a gap between then.

Some discussion on this:
https://kotlinlang.slack.com/archives/CJLTWPH7S/p1724421782507849