Split notification channels from the senders
We need to allow ourselves to have notification senders which do not
uniquely own their notification channel. To do this, extract the
notification channel part as a separate interface, with its own separate
implementations

This also allows us to defer the creation of the notification channels
to when we actually want to send the notification rather than all at app
startup.