Start potentially supporting partial gql responses
Use iorNel which instead of `Either<Error, Data>``
returns `Ior<Nel<Error>, Data>`
Ior is basically an `Either` which can have *both* left and right at
the same time, which is exactly what we want when we get partial
responses.

This is just done internally, but everything is still returned as
Either, to keep the app running just as it used to. We can consider
using this at a later point, this is just to make it possible.

This is for problems like when we've had the entire home screen error
out on some members when one field failed, when we could've just had
that one field be null and still showed all the rest of the data

See also articles like:
https://adambennett.dev/2024/09/using-feature-flags-in-graphql/
About how such partial responses break features and some other options
for working around it. While the @include directive works for what they
show there, just being able to work with partial responses properly is
the more solid long-term approach without having to add @include
directives everywhere.