Make ValidatedInput have an explicit Output type
This way we can ensure that a given input can be a specific type on the
output side. This can often mean that the input is nullable, but the
output is valid only when it's not null

Also turn it into an interface to make it simpler to reason about the
public API it exposes.

Add a `NoopValidator` as an easy to add implementation for compose
@Previews