Transform the content into validContent before submitting
Use the new ValidatedInput API to get one valid output immediately
through one validator. Use this output as input for the ValidContent
object which should only exist if everything was as expected.
Take this output as the input for the next step, and ensure that we
store this data back into the persistent moving flow storage so that
it's still there on the next step or if we go forward and back in the
flow again.