Make ComparisonTable and QuoteCard more robust to text size changes
Ensure that the size of the content inside a cell in the
ComparisonTable and QuoteCard is independent of the text size.
This is done by placing an invisible "H" in the background and making
sure the actual content is at least as large as that "H".

Additionally, the width of the header cells in the ComparisonTable is
now fixed to the widest header text, ensuring consistent layout.

The `Cell` composable is used as the right container for all the items
that should be considered a "cell", as it sets up the right content size
automatically, and makes sure that the content's min constraints are
fixed to be as big as the cell itself is.