Improvements on LayoutWithoutPlacement
Ensure that the `content` child gets the minimum constraints setup to be
at least as big as the `sizeAdjustingContent` was, but also allow itself
to adjust the parent's size, sometihng which `matchParentSize` does not
let you do.
This as a consequence removes the `BoxScope` from the API, which
requires a couple of call-site changes, but nothing major.