Introduce OverlaidIndication to handle the full row indication logic
overlaidIndicationConnection serves as the modifier to be added to the
places where the source of the indication comes from, which syncs all
interactions back to the interaction source which is used in the
OverlaidIndication.
The normal `clickable` modifier is then used in the rows since that one
correctly passes all interaction events, including hover events, long
clicks etc, which are otherwise harder to do manually with a
pointerInput

I basically also used this as an opportunity to learn how to work with
custom modifiers. I followed the docs here if you're curious:
https://developer.android.com/develop/ui/compose/custom-modifiers#implement_a_custom_modifier_using_modifiernode