Move the sheet state outside of the VM
Use HedvigBottomSheetState to store the information of the sheet.
Dismiss it as normal when clicking "close" or when clicking outside.
Also hook into the state of `showReferralCodeSuccessfullyChangedMessage`
to dismiss it when it turns true, since we know that we have
successfully submitted a code then.

At the same time, remove the extra horizontal padding since the bottom
sheet has a default of 16.dp horizontal which is what we wanted here.