Make sure that modal bottom sheets don't give focus on appearance
Due to how he modal sheet is implemented
https://github.com/oleksandrbalan/modalsheet/blob/519938d25af1b9f41a409c5514f36a6cce0f4cb0/modalsheet/src/main/java/eu/wewox/modalsheet/FullscreenPopup.kt#L135
[ModalSheet] automatically requests focus on appearance. This means that if there is a focusable child available,
it gets focus immediately. This commonly is a TextField, which also brings the keyboard up. The keyboard ends up
coming up way too early, before the sheet manages to show, which results in a very awkward and janky animation
where the keyboard and the sheet are coming up in different timings, hiding each other while the insets are not
quick enough to catch up to make them go up in sync.
Making this box focusable means that it itself grabs the focus instead, not showing the keyboard on appearance.