Remove overload with a built-in bottom button
This ended up making it harder to work with insets, since any insets
added on the content slot would end up showing *above* that button,
which is not what we expect here.
Remove the `sheetPadding` parameter for everyone, since we do not want
to inset the sheet itself anything other than around the top and
horizontally, as we want the sheet to always go all the way to the
bottom of the screen.

Update all callers accordingly