Stop scrolling to the bottom of the chat on the first frame
We simply drop the first ever "latest message" as it is the one which
appears on the first time any messages are rendered on the screen.
In the cases where we just came to this screen, we will have been at the
beginning of the list regardless. And in the scenarios where we come
back to this screen which has existed before we instead want to stay at
the scroll position we had in the past