Add stricter input controls in contact info screen
Be more careful regarding cleaning up the input coming from the backend
where phone numbers can have spaces in them.
Make the text fields themselves try to clean up the input by using the
new InputTransformation from the BasicTextField2 overload which takes
in a TextFieldState.
Make use of that TextFieldState as the source of truth of