Fix bottom sheet showing before properly navigating back
Now it will wait until we've settled back to the home screen before it
considers showing the bottom sheet.
There are no changes required as to when we mark the cross sells as
seen, since that step is linked with the sheet actually being actually
shown, which now simply happens a bit later.