Make bottom sheet animate both in and out
Make isVisible a local mutableState which controls whether the sheet
should be in composition or not.
When the bottom sheet itself enters composition, it animates in by
itself.
When the sheet has finished animating out,
rememberInternalHedvigBottomSheetState takes care of turning isVisible
as false.
Had to remove the scroll to bottom part as it was interfering with the
scrollable state of the sheet's content.