Add dialog style for titled dialogs
This is done to allow a centralized spot of adding a title and making
sure it's read as a heading for a11y purposes

Introduce a new `DialogStyle.TitleNoButtons` to support dialogs with a title but no buttons.
Update existing dialog usages (e.g., `RadioSelectDialog`, `CheckboxSelectDialog`) to utilize the new `DialogStyle` options.
Added semantics for the dialog title to be recognized as a heading for accessibility.