Adjust ExpandablePlusCard to have the description as a separate node
By letting the clickable modifier merge its descendants, the title
is read along with the stateDescription added on the card itself.
The expandable description is then made to be its own merged
descendants parent, which makes it so that everything in there is read
separately from the entire clickable card.