Refactor CrossSells BottomSheet
Add a `sheetPadding` parameter to `HedvigBottomSheet` to allow for
padding of the sheet itself.
Introduce a `BottomSheetStyle` to control properties like transparent
background.
Add a `cornerXLargeBottom` shape to make the content match the shape
above it which is already curved at the top
Update `CrossSellsSheetContent` to make use of this new style, and add
add the necessary background colors only to the right parts of the UI,
while keeping the rest with a trasnparent background to give the
illusion that the content is floating above the rest.
No way to adjust the background scrim to change it into a blur, so that
part of the design is skipped