Refactor ApolloOperationError and logging
This commit refactors `ApolloOperationError` by moving it to its own
module `apollo-operation-error`.
This allows it to be used from inside the logging-public module.
This gives us the freedom to adjust the logs that come from apollo
operation errors.
The `logcat` function now accepts an `ApolloOperationError` parameter,
allowing for more specific logging based on the error type. The log
priority is adjusted to `WARN` if the error is due to an unauthenticated
user, preventing excessive error reporting for such cases.

The `OperationResult` sealed interface was deleted as it was only used
in legacy code