Workaround the a11y problem introduced by the workaround on textfield
The item itself can't be clickable since the textfield focusable
modifier eats up all interactions. Yet we can't rely on that from a11y
due to the text field looking read-only.
This is a workaround to the workaround which may perhaps warrant a
completely new component or special handling inside the text field
component itself.