Fix a crash with a skiko file not found
Apparently, when the android target is not specified in a project using
CMP, it sometimes is possible that when trying to use things that have
a different implementation on Android it will instead try to use the jvm
option which is not present on Android.

Adding the android target means that it now knows to pick the correct
actual implementation which works as intended.